Troublehooting type stripping issue
